Fix name of the experience package and move adding points into util
