+++ src/main/java/com/eternalcode/discordapp/experience/Experience.java
@@ -1,4 +1,4 @@
-package com.eternalcode.discordapp.expierience;
+package com.eternalcode.discordapp.experience;
 
 public class Experience {
